start working on revamped shooter cmds

+++ src/main/java/frc/robot/subsystems/AngleAdjustSubsystem.java
@@ -37,7 +37,8 @@ public class AngleAdjustSubsystem extends SubsystemBase {
                     absoluteEncoder.getAbsoluteRotations()));
     }
 
-    public boolean atDesiredAngle(double desired, int tolerance) {
+    public boolean atDesiredAngle(double desired,
+                                  double tolerance) {
         return (MotorUtil.inTolerance(desired, getAngle(), tolerance));
     }
